feat(extensions): Profile Editor UI, CRX import, tests, README
Day 3-4 of feature-extensions plan:

UI (Profile Editor):
- Add 'Engine' dropdown listing bundled Chromium-for-Testing + auto-detected Chrome / Brave / Edge / Vivaldi / Opera / Chromium installs
- Add 'Extensions' panel with Add folder / Import .crx buttons, per-row enable toggle and remove
- ExtensionItemViewModel binds to underlying ProfileExtension model
- Save: persist EnginePath + diff extension list (insert new, update existing, delete removed)
- Window grew to 680x740 to fit the extra controls

CRX import:
- Use Avalonia StorageProvider for native folder/file pickers
- CrxImporter.Extract auto-extracts to %LOCALAPPDATA%/ZeroBrowser/extensions/<profile>/<random>
- Validate manifest.json after extract, surface errors via ExtensionStatusMessage
- ReadManifestName auto-labels extensions (handles __MSG_*__ i18n placeholders)

Tests:
- CrxImporterTests: extract v2 + v3 synthetic CRX, manifest validation, name parsing, MSG placeholder stripping, bad-magic / unsupported-version / missing-manifest error paths (9 tests)
- BrowserDetectorTests: API safety, file existence, uniqueness (3 tests)
- ProfileExtensionRepositoryTests: insert+list, update toggle/rename/sort, delete, FK cascade-delete on profile removal, sort_order ordering, EnginePath round-trip, EnginePath clearing via update (7 tests)

Bug fix:
- Dapper materialization needs Int64 column types for INTEGER columns; ExtensionRow.sort_order changed from int to long; cast back to int in MapExt

README:
- New 'Extensions (per profile)' section explaining engine modes (deterministic Chromium-for-Testing vs branded user-installed) and the anti-detect tradeoff

Test results: 67/67 (was 48 + 19 new). 0 warnings, 0 errors.

+++ src/ZeroBrowser.Storage/Sqlite/ProfileRepository.cs
@@ -185,7 +185,7 @@ public sealed class ProfileRepository
         Name      = row.name,
         Path      = row.path,
         Enabled   = row.enabled != 0,
-        SortOrder = row.sort_order,
+        SortOrder = (int)row.sort_order,
         CreatedAt = DateTimeOffset.FromUnixTimeSeconds(row.created_at)
     };
 
@@ -196,5 +196,5 @@ public sealed class ProfileRepository
 
     private sealed record ExtensionRow(
         string id, string profile_id, string name, string path,
-        long enabled, int sort_order, long created_at);
+        long enabled, long sort_order, long created_at);
 }
